pipeline/requirements.txt
# ML Pipeline Service Dependencies
# Core ML/Audio Processing
torch>=2.6.0
torchaudio>=2.6.0
transformers>=4.38.0
librosa>=0.10.0
soundfile>=0.12.0

# Feature Processing & Search
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.8.0
# faiss-gpu>=1.8.0  # Use for GPU deployment

# Data Processing
pandas>=2.0.0
scipy>=1.10.0

# API Framework (for pipeline service)
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0

# Development/Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0